+++ packages/google/src/google-generative-ai-options.ts
@@ -37,11 +37,12 @@ export type GoogleGenerativeAIModelId =
   | 'gemini-2.5-flash-native-audio-preview-09-2025'
   | 'gemini-2.5-flash-native-audio-preview-12-2025'
   | 'gemini-2.5-computer-use-preview-10-2025'
-  | 'gemini-3.1-pro-preview'
-  | 'gemini-3.1-pro-preview-customtools'
   | 'gemini-3-pro-preview'
   | 'gemini-3-pro-image-preview'
   | 'gemini-3-flash-preview'
+  | 'gemini-3.1-pro-preview'
+  | 'gemini-3.1-pro-preview-customtools'
+  | 'gemini-3.1-flash-image-preview'
   // latest version
   // https://ai.google.dev/gemini-api/docs/models#latest
   | 'gemini-pro-latest'
@@ -180,9 +181,13 @@ export const googleLanguageModelOptions = lazySchema(() =>
               '9:16',
               '16:9',
               '21:9',
+              '1:8',
+              '8:1',
+              '1:4',
+              '4:1',
             ])
             .optional(),
-          imageSize: z.enum(['1K', '2K', '4K']).optional(),
+          imageSize: z.enum(['1K', '2K', '4K', '512']).optional(),
         })
         .optional(),