+++ packages/react/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @ai-sdk/react
 
+## 3.0.108
+
+### Patch Changes
+
+- ai@6.0.106
+
+## 3.0.107
+
+### Patch Changes
+
+- Updated dependencies [58bc42d]
+  - @ai-sdk/provider-utils@4.0.16
+  - ai@6.0.105
+
+## 3.0.106
+
+### Patch Changes
+
+- ai@6.0.104
+
+## 3.0.105
+
+### Patch Changes
+
+- ai@6.0.103
+
+## 3.0.104
+
+### Patch Changes
+
+- ai@6.0.102
+
 ## 3.0.103
 
 ### Patch Changes